requirements.txt
python-docx==0.8.11
docx2txt==0.8
doc2text==0.2.4
pyinstaller==6.3.0